change: examples: serve the V1 network over hyper, drop actix
# Summary

`network-v1-http` now ships a `Server` for the inbound half of the legacy
V1 network, replacing the optional Actix handler module. The crate becomes
a self-contained, single-stack mirror of `network-v2-http`.

# Details

The crate previously implemented only the outbound half; inbound RPCs were
an `actix` feature that nothing in the workspace enabled, so the V1 network
had no runnable server and no test coverage. `Server` closes that gap and
lets an example app run entirely on V1.

Actix was the odd stack out: `app-http` and `network-v2-http` are both
hyper-based, so a V1 app built on the old module would have had to link two
HTTP servers. Mirroring `network-v2-http/src/server.rs` instead keeps a V1
app one dependency away from its V2 counterpart, which is what makes the
two comparable as examples.

`/snapshot` decodes the chunked `InstallSnapshotRequest` and forwards it to
`ChunkedSnapshotReceiver::install_snapshot()`, so the receiving end needs a
file-like snapshot to assemble chunks into. That is expressed as a bound on
the state machine's `SnapshotData` rather than pinning a concrete type.

Every route answers with a serialized `Result<Resp, Err>` that splits the
API error out of `RaftError`, matching the shape the existing `Network`
client already decodes; the wire format is unchanged.

The crate is split into `client.rs` and `server.rs` following the
one-main-type-per-file convention and `network-v2-http`'s layout.

Upgrade tip:

The `actix` feature is gone. Replace

    actix_web::App::new()
        .app_data(Data::new(raft.clone()))
        .configure(network_v1_http::actix::configure::<C, SM>);

with

    network_v1_http::Server::new(raft.clone()).run(raft_addr).await?;
