+++ examples/network-v1-http/Cargo.toml
@@ -18,15 +18,19 @@ repository = "https://github.com/databendlabs/openraft"
 openraft = { path = "../../openraft", features = ["serde", "type-alias"] }
 openraft-legacy = { path = "../../legacy", features = ["serde"] }
 
-actix-web = { version = "4.0.0-rc.2", optional = true }
-reqwest    = { version = "0.12.5", features = ["json"] }
-serde      = { version = "1.0.114", features = ["derive"] }
-tokio      = { version = "1.35.1", features = ["full"] }
-tracing    = { version = "0.1.40" }
+bytes          = { version = "1.0" }
+http-body-util = { version = "0.1" }
+hyper          = { version = "1.0", features = ["http1", "server"] }
+hyper-util     = { version = "0.1", features = ["tokio"] }
+log            = { version = "0.4" }
+reqwest        = { version = "0.12.5", features = ["json"] }
+serde          = { version = "1.0.114", features = ["derive"] }
+serde_json     = { version = "1.0.57" }
+tokio          = { version = "1.35.1", features = ["full"] }
+tracing        = { version = "0.1.40" }
 
 [features]
 default = []
-actix = ["dep:actix-web"]
 
 [package.metadata.docs.rs]
 all-features = true
